chore: refactor lambda function code
