+++ src/funcs/cost-reporter.lambda.ts
@@ -1,36 +1,17 @@
-import {
-  CostExplorerClient,
-  GetCostAndUsageCommand,
-  GetCostAndUsageCommandInput,
-  GetCostAndUsageCommandOutput,
-} from '@aws-sdk/client-cost-explorer';
+import { CostExplorerClient } from '@aws-sdk/client-cost-explorer';
 import { IncomingWebhook } from '@slack/webhook';
 import { Context } from 'aws-lambda';
+import { GetServiceBilling, GetTotalBilling } from './lib/get-billing-command';
+import { GetDateRange } from './lib/get-date-range';
 
-export class EnvironmentVariableError extends Error {
+export class MissingEnvironmentVariableError extends Error {
   constructor(message: string) {
     super(message);
-    this.name = 'EnvironmentVariableError';
+    this.name = 'MissingEnvironmentVariableError';
   }
 }
 
 export interface EventInput {
-}
-
-interface DateRange {
-  readonly start: string;
-  readonly end: string;
-}
-
-export interface TotalCost {
-  readonly unit: string;
-  readonly amount: number;
-}
-
-export interface ServiceCost {
-  readonly service: string;
-  readonly unit: string;
-  readonly amount: number;
 }
 
 const ceClient = new CostExplorerClient({
@@ -43,108 +24,21 @@ export const handler = async (event: EventInput, context: Context): Promise<stri
 
   // do validation
   if (!process.env.SLACK_WEBHOOK_URL) {
-    throw new EnvironmentVariableError('SLACK_WEBHOOK_URL environment variable not set.');
+    throw new MissingEnvironmentVariableError('missing environment variable SLACK_WEBHOOK_URL en.');
   }
   if (!process.env.SLACK_POST_CHANNEL) {
-    throw new EnvironmentVariableError('SLACK_POST_CHANNEL environment variable not set.');
+    throw new MissingEnvironmentVariableError('missing environment variable SLACK_POST_CHANNEL environment variable not set.');
   }
 
   // 👇Calculate Date Range
-  const dateRange: DateRange = (() => {
-    const dateFormatString = (date: Date): string => {
-      return (date.getFullYear()) + '-' + ('00' + (date.getMonth() + 1)).slice(-2) + '-' + ('00' + (date.getDate())).slice(-2);
-    };
-
-    const now = new Date(Date.now());
-    if (now.getDate() === 1) {
-      // Last month
-      return {
-        start: dateFormatString(new Date(now.getFullYear(), now.getMonth() - 1, 1)),
-        end: dateFormatString(new Date(now.getFullYear(), now.getMonth(), 0)),
-      };
-    }
-
-    return {
-      start: dateFormatString(new Date(now.getFullYear(), now.getMonth(), 1)),
-      end: dateFormatString(new Date(now.getFullYear(), now.getMonth(), now.getDate() - 1)),
-    };
-  })();
+  const dateRange = new GetDateRange();
   console.log(`DateRange::${JSON.stringify(dateRange, null, 2)}`);
 
   // 👇Get Total Billing
-  const totalBilling: TotalCost | undefined = await (async () => {
-    const input: GetCostAndUsageCommandInput = {
-      TimePeriod: {
-        Start: dateRange.start,
-        End: dateRange.end,
-      },
-      Granularity: 'MONTHLY',
-      Metrics: [
-        'AMORTIZED_COST',
-      ],
-    };
-    console.log(`TotalBilling:Command:Input:${JSON.stringify(input)}`);
-    return ceClient.send(new GetCostAndUsageCommand(input))
-      .then((data: GetCostAndUsageCommandOutput) => {
-        if (data && data.ResultsByTime && data.ResultsByTime.length === 1) {
-          const cost = Object(data.ResultsByTime[0]).Total.AmortizedCost;
-          const result: TotalCost = {
-            unit: cost.Unit,
-            amount: cost.Amount,
-          };
-          console.log(`TotalBilling:Command:Output(Shaped):${JSON.stringify(result)}`);
-          return result;
-        }
-        return undefined;
-      })
-      .catch((error) => {
-        console.log('Error caught...');
-        console.log(`Error:${JSON.stringify(error)}`);
-        return undefined;
-      });
-  })();
+  const totalBilling = await (new GetTotalBilling(ceClient)).execute(dateRange);
 
   // 👇Get Service Billings
-  const serviceBillings: ServiceCost[] | undefined = await (async () => {
-    const input: GetCostAndUsageCommandInput = {
-      TimePeriod: {
-        Start: dateRange.start,
-        End: dateRange.end,
-      },
-      Granularity: 'MONTHLY',
-      Metrics: [
-        'AMORTIZED_COST',
-      ],
-      GroupBy: [
-        {
-          Type: 'DIMENSION',
-          Key: 'SERVICE',
-        },
-      ],
-    };
-    console.log(`ServiceBillings:Command:Input:${JSON.stringify(input)}`);
-    return ceClient.send(new GetCostAndUsageCommand(input))
-      .then((data) => {
-        const billings: ServiceCost[] = [];
-        if (data.ResultsByTime && data.ResultsByTime.length === 1) {
-          for (const item of Object(data.ResultsByTime[0]).Groups) {
-            billings.push({
-              service: item.Keys[0],
-              unit: item.Metrics.AmortizedCost.Unit,
-              amount: item.Metrics.AmortizedCost.Amount,
-            });
-          }
-          console.log(`ServiceBillings:Command:Output(Shaped):${JSON.stringify(billings)}`);
-          return billings;
-        }
-        return undefined;
-      })
-      .catch((error) => {
-        console.log('Error caught...');
-        console.log(`Error:${JSON.stringify(error)}`);
-        return undefined;
-      });
-  })();
+  const serviceBillings = await (new GetServiceBilling(ceClient)).execute(dateRange);
 
   console.log(`TotalBilling: ${JSON.stringify(totalBilling, null, 2)}`);
   console.log(`ServiceBilling: ${JSON.stringify(serviceBillings, null, 2)}`);
